Refactor interfaces
rename and include them in one file
