--- client/IClient.go
@@ -1,7 +1,0 @@
-package client
-
-import receipt "subscribator-go/client/receipt"
-
-type IClient interface {
-	ValidateTransaction(transaction string) receipt.IReceipt
-}
