--- client/receipt/IReceipt.go
@@ -1,7 +1,0 @@
-package client
-
-type IReceipt interface {
-	GetReceipt() string
-	GetTransaction() string
-	GetProvider() string
-}
