--- client/transaction/ITransaction.go
@@ -1,6 +1,0 @@
-package client
-
-type ITransaction interface {
-	GetTransaction() string
-	GetProvider() string
-}
