+++ test/client_test.go
@@ -3,14 +3,13 @@ package test
 import (
 	"fmt"
 	client "subscribator-go/client"
-	transaction "subscribator-go/client/transaction"
 	"testing"
 )
 
 func TestAppleClient(t *testing.T) {
 
 	itunesClient := client.ItunesClient{}
-	itunesTransaction := transaction.NewItunesTransaction("token_itunes")
+	itunesTransaction := client.NewItunesTransaction("token_itunes")
 
 	var receipt = itunesClient.ValidateTransaction(itunesTransaction.GetTransaction())
